Add support for determining locol vs remote change
Add the ability for gps to know if your local patch has changed since
the last recorded operation or if the remote patch has changed since the
last operation or if both have changed since this last operation. This
also now categorizes the remote branch having multiple commits as the
remote patch having changed instead of it short-circuiting the list
command. This also means that the representation of the states in the
list command are different.

The approach I took was to change our state tracking and storage so that
with all of the operations (branch creation, syncing, and request
review) the patch diff hash at that point in time is recorded along with
the state. This allows me to get the local patch's diff hash and the
remote patches diff hash and compare them to the diff hash recorded with
the last operation. This enables us to determine if a change has
happened to the local patch, to the remote patch, or to both.

This relates to issue #71.

[changelog]
changed: state storage format to support local & remote change detection
added: local patch change detection & presentation to the list command
added: remote patch change detection & presentation to the list command

ps-id: 4c59824b-5b7d-4818-96af-452c4502a418