Add support for communal repository hooks
Add support for communal repository hooks so that teams can enforce hook
implementations within a repository. I also cleaned up the README hooks
section to try and make the it clear how to get going with hooks.

The approach I took was simply to add checking for communal repository
hooks first and then to fall back to the repository specific personal
hooks and finally to the global user specific hooks.

This relates to issue #73 and #70.

[changelog]
added: support for communal repository hooks

ps-id: d3d61fdb-e08d-470c-a6fa-577fca80560d