Add color & human readable err output to rr cmd
The intention with this change is to make things clearer when the
request review command fails. Prior to this change the output was kinda
just like any other output in the sea of output. Hopefully this change
pulls the error out of the sea a bit by adding the blank lines around it
as well as adding the Red coloring of the error message. I also adding
transaltion from the error types to human readable messages so that
users might better understand what happened.

This is related to issue #80.

[changelog]
added: colored & human redable err output to request review command

ps-id: 59812e17-fd6a-41dc-b364-2332ea3a72e3