Add request_review_post_sync hook messaging
Add specific messaging around the request_review_post_sync hook as it is
a blocker for the requset-review command to function. So I wanted to
make it clear to the user and give them information about it and where
they can go to find more information. I also wanted to make the hook
found but not executable case well known and how to address it easy for
the user as it as well is blocker for the request-review command
functioning properly.

This relates it issue #80.

[changelog]
added: request_review_post_sync messaging in requset-review command

ps-id: 81428726-0981-41cb-9318-cf76eb849edc