Trim whitespace off GPG signature
Modify the gpg_sign_string() function so that it will trim trailing
whitespace off the end of the signature. I am doing this because as part
of issue #86's investigation I discovered that there was an extra `0a
20` as hex characters after the end of the GPG signature that isn't
there when using git proper commands. Adding trim() corrects it so that
the resulting output matches the behavior of git itself. My hope was
that this would also fix the problem discussed in issue #86 but it does
not. Despite that I believe this is still a worthy fix.

Related to issue #86.

[changelog]
fixed: problem where GPG signatures had an extra `0a 20` hex characters