Fix cherry pick to use configured committer
Previously the cherry pick implementations where just forwarding on the
committer from the original commit. This isn't really accurate as the
commit created wasn't committed by the committer of the commit being
cherry-picked. Forwarding of this committer also had the side effect of
forwarding the timestamp of the commiter which also needed to be
corrected to be the timestamp of when the cherry pick occured.

This issue was discovered while digging into issue #86 but does not
resolve issue #86.

My approach was to simply use repo.signature() as it gets the configured
user from the git config and use the current time to create a signature
object.

This is related to issue #86.

[changelog]
fixed: cherry pick committer & timestamp forwarding