Add state to `list` command to indicate if behind
Basically there a patch is behind if the request-review branch
associated to that patch was based on a previous state of its base which
has since changed. I added a way to represent this in the `list` command
so that users would be aware of this and could address it by either by
running a `sync` or `request-review` command again for that patch. This
is also important because I am planning to change the `integrate`
command back to using the remote request-review branch for the
integration and not overwriting the local branch on integrate as using
the local branch is causing GitHub, Bitbucket, etc. to not identify
integrations as "Merged" and instead leaving them open or maybe once the
branches are deleted then identifying them as "Closed".

This relates to issue #86.

[changelog]
added: concept of patch being behind to `list` command

ps-id: 20da9c9d-a198-462c-b215-55afb0eabea8