Fix logic of git::singular_commit_of_branch()
The git::singular_commit_of_branch() function was previously incorrectly
implemented to find the common ancestor between the branch and HEAD
(where the user is currently checked out on). This is incorrect and will
often result in getting a singular commit failing because it thinks
there are multiple commits in the branch when there are not actually
multiple commits in the branch.

The approach I took to correct this was to remove the getting HEAD and
finding the common ancestor between the branch and HEAD and instead
requiring a base_oid be passed in that we can then use to compare the
branch too.

[changelog]
fixed: bug where patches were incorrectly identified as having remote changes