Move commit_is_behind out to its own module
My intention was simply to move the commit_is_behind() out of the list
module and into its own module that can be shared across other modules.
Specifically I am planning on using it in the `integrate` implementation
so that I can check if the patch is behind prior to moving forward with
the integration. I also renamed it just to clean up the naming a bit.

This relates to issue #86.