Add check for patch being behind to integrate
Add check for the patch being behind to the integrate command
implementation such that if the patch is behind then it will abort the
integrate command with an error message explaining what is going on.
This also means that our integration approach will differ now. Instead
of using the local branch that is automatically rebased it is going to
integrate from the remote request review branch. This helps make sure
that GitHub/Bitbucket/etc detect the integration as a merge.

The approach I took was using the commit_is_behind module to check if
the commit attempting to be integrated was behind in it's request-review
branch. If it was then I short-circuit with an error indicating that the
patch was behind. Otherwise I allow it to progress forward. I removed
the code that reset the local branch as we no longer want to do that.
Instead we want to use the remote request review branch so I modified
the push to actually push from the remote request review branch. Then in
the command layer I check for the specific error type and display an
error message to the user in that case.

This relates to issue #86.

[changelog]
fixed: issue where integrations weren't identified as merges by Bitbuckt/GitHub