Extract patch_status() to ps/private module
Extract the patch_status() function and it's associated public interface
types to its own module nested in ps/private so that we will be able to
use the patch_status() functionality in multiple places. For example I
am planning on using in the `pull` command implementation in the future
to automate the updating of the request-review branches of patches when
it is safe to do so without losing changes.

The approach I took to accomplish this was simply to extract the
patch_status() function out into a new module including all it's public
interface types, add that module to the ps/private module, and then
change the existing references to use this new module.

This is related to issue #87.

ps-id: 8ed5e04f-6bed-4c9c-adea-142843fdaae1