Add check for upgrade to list command
Add check for upgrade to the tail end of the list command so that people
can easily find out when there is a new release of gps.

The approach I took with this was basically fetching the latest release
from the GitHub repository and comparing that version to the version in
the running app. If that version is newer then notify the user about the
new version. If it is older or fails for some reason just quietly do
nothing.

This relates to issue #90.

[changelog]
added: check for upgrade to tail end of list command