Fix ! showing when patch changed due to cherry-pick
There is a bug currently where the ! state shows when you have
synced or requested review and there are patches below it in the stack
that alter the same file such that when the cherry-pick occurs onto the
base of the stack the actual patch changes in terms of line numbers and
context. My intention is for this change to resolve that situation by on
operations not only recording the patch diff id of the commit in the
patch stack but also the patch diff id of the successfully cherry-picked
commit. This should provide a reference to compare the computed remote
against to see if the remote has changed since the last operation.

Given that this changes how state is stored I initially changed the path
to the state store file so that we wouldn't have any conflicts. Then I
proceeded to modify the state_Management module such that it would
include the remote commit's diff id in both the sync and requested
review states. Then I update sync and request_review commands to include
that value in the state that is stored when those operations occur.
Finally I updated the patch_status module and how we determine the state
which in turn controls the ! character such that remote difference is
determined by comparing against the operational remote diff id now
instead of the local diff id.

Note: This is a BREAKING change as it uses a different state store so
when a user upgrades to this it will act as if they have no state.

This is related to issue #88.

[changelog]
fixed: ! state modifier appearing when remote hadn't actually changed

ps-id: f3e19685-294f-4768-91f4-b1496b13b9aa