Move new release check to a thread
Move new release check to a thread so that it is working while we are
off actually performing the list command. Then when the list command is
complete we join the thread, get the return value, and check if we need
to notify the user or not. Hopefully this generally means the check for
a new release is completed by the time the list command is done and it
can just get the value. If however it isn't complete then we block
waiting for it to finish.

The approach I took was modifying the notify_of_newer_release() function
to take in the argument rather than calling it directly. This allows us
to split the fetching out into a spawned thread while leaving the
notifying on the main thread.

[changelog]
changed: new release check to a background thread

ps-id: 7211d133-c12c-4037-837a-0fdec3342f95