Remove check for release from list command
Remove check for new release from the list command because it is a
command that should be as fast as possible, is used all the time, and
inately doesn't need network access to work. Therefore, it seems ill
fitting to have it make a network request to check if there is a new
release available.

It makes more sense to have a command like `pull` which you expect to
make a network request do the check instead. Also it is a command that
generally takes longer so it gives it more time to perform the check in
parallel without impacting the perceived performance of the command.

[changelog]
removed: new release check from the `list` command

ps-id: 2d6ac1fb-a8bd-43ce-a6fe-62053b5e1a14