Add new release check to the `pull` command
Add new release check to the `pull` command since I previously removed
it from the `list` command as it didn't seem fitting there and it could
impact perceived performance there. Here it is less likely to impact
perceived performance and it is clear from the consumer standpoint that
the `pull` command is going to reachout and make network requests so it
makes more sense to be aligned with this command.

I added it the same way that we had it in the list command with one
exception. In this implementation if the check for the new release fails
for any reason we now quietly swallow the error as it isn't crucial to
functionality of the command and to the user it is just noise they don't
want to deal with or think about. I am not sure how we find out if it
stops working other than someone discovering that a release was made and
that they weren't notified. But maybe that is fine.

[changelog]
added: new release check to the `pull` command

ps-id: 1e540f16-3e30-4c50-a7f4-59072054ba95