Fix gps int exit w/ ShowFailed(ExitSignal(13))
There is currently an issue with gps int where it will exit with the
ShowFailed(ExitSignal(13)) when the prompt_for_reassurance option is
enabled and you are attempting to integrate a large patch and you quit
out of the pager showing the changes prior to traversing to the last
page.

Turns out this is due to the fact that when we do the show::show()
command it is actually running `git show --pretty=raw` for the patch
which pipes its output to the pager configured either with Git or with
the shell. Pagers have a limited buffer they use to store data so when
large amounts of data are sent to them they only accept up to that
buffer amount until the user traverses through the pages at which point
it throws away the old data and allows the new data to be written to the
pipe. If the pager is exited prior to traversing to the last page it
will cause a SIGPIPE for Git because all of git shows output hasn't been
written to the pipe yet.

To address this I decided it made the most sense to continue to use the
default behavior with a pager as that is what users would want and
expect. But to check for this explicit error and display a warning in
this case educating the user as to what is likely going on but still
allowing the gps int process to continue forward.

I briefly contemplated swallowing the error all together. But in general
that is a bad idea as there are likely other ways that the SIGPIPE
failure could occure.

This resolves issue #120.

[changelog]
fixed: gps int exiting with ShowFailed(ExitSignal(13)) for large patches

ps-id: 3114260d-2b3e-4478-bf3b-90619f1208f7