Update home-dir to get updated version of nix
Update home-dir to get updated version of nix because the previous
version of nix we were depending on had some security vulnerabilities we
wanted to eliminate.

To do this I forked the home-dir repository and pushed up a branch that
upgrades to the latest version of nix. I then opened a pull request from
this branch to the upstream repository. However, I didn't want to wait
for this update. Therefore, I use the [patch.crates-io] to override the
the home-dir repository and branch to use the patched version.

In order to get this to work I also had to upgrade libc to a version
that would meet all the requirements as the version prior to this no
longer met the requirements of the latest version of nix.

[changelog]
fixed: vulnerable dependency, nix via home-dir

ps-id: 31aee10f-d783-4494-84ca-33aeb4224a11