Make cherry_pick_no_working_copy take time offset
Make cherry_pick_no_working_copy take time offset so that the timestamp
(seconds since epoch) associated with the committer can be offset by a
value in certain scenarios. Specifically this is useful in scenarios
where you need to do multiple cherry picks of the same commit within a
single second. This for example currently happens as part of the
request_review_branch() process when a commit has no patch stack id.
That is because it first does a cherry-picking process to amend the
commit with the patch stack id and then immediately after it does the
cherry-picking process to create the request review branch and cherry
pick the commit into it. If all that happens in the same second the SHA
of both cherry picked commits will be the same and the ps/rr/ branch and
the patch on the stack will be the same and both references will point
to it. This will cause patches in that state to automatically collapse
away when you do a gps int which can be confusing to the user.

Adding this ability to the the cherry_pick_no_working_copy() function
should enable us to in specific cases like the case above offset the
timestamp say by 1 second in turn causing the content of the commit to
be different, in turn causing the SHA of the commit to be different, in
turn preventing the scenario I described above.

This is related to issue #78.