Make request_review_branch() offset rr cherry-pick
Make request_review_branch() offset the rr branch cherry-pick by 1
second so that it doesn't have the same commit timestamp (seconds since
epoch) as the add patch stack id process does. This will cause it to be
a separate Git commit in the Git tree which in turn will make it behave
as all the other scenarios behave. Effectively consistently behave now.

My approach was simply to change the offset from 0 to 1 in this
particular scenario. This was only possible thanks to a previous patch
where you added the ability to specify an offset in the cherry pick
process.

This resolves issue #78.

[changelog]
fixed: gps int would sometimes collapse away integrated patches (#78)