Upgrade dependencies
My goal with this was to upgrade all of the dependencies we use so that
we wouldn't fall behind.

This involved only making one code changes which was to switch from
using the `to_hyphenated()` method on Uuid to the `hyphenated()` method.
Everything else had no build issues and as far as my testing has gone
has no runtime issues I am aware of.

[changelog]
changed: to using latest versions of dependencies