Fix bug where sync wasn't updating rr state
There is a bug where if you have a patch that was requested review, then
you modify the patch in your patch stack, and then sync that patch it
shows a status of +! incorrectly. It seems this happens because the sync
command currently doesn't update the stored state when the patch being
synced was currently in the requested review state.

To correct this I modified the sync command so that when the patch is in
a requested review state it will correctly update the state.

This resolves issue #122.

[changelog]
fixed: requested review patch status not updating on sync

ps-id: 7554734d-3d66-4c44-ba92-7288f6edc93c