pull after integrate
If the user configured gps to pull latest after integrating a patch, pull down latest

This is done to respect the configuration the user has

I added a call to the public pull interface after integrate is complete. I called the pull method instead of the pull command to avoid checking for a new gps version at the end of the pull. If the pull returns an error, simply warn the user and complete the integrate action.

[changelog]
added: pulling after integrating with user configuration

ps-id: ccf26a8d-e68a-46d7-9a81-0d2d09d4ae8d