Add config to pull after integrate
Allow user to specify whether or not they want gps to pull latest after they integrate a patch

This is done to give our users the ability to pull latest after integrating a patch, avoiding having to do it manually because integrating always leaves them in a diverged state.

I added the pull_after_integrate config to the integrate portion of the configs. I defaulted it to `false` to maintain current behavior.

[changelog]
added: configuration option to pull after integration