Add upstream-patches command
Add upstream-patches (up) command so that users would be able to quickly
see what patches are upstream and would be pulled in when they run a
`gps pull`.

Need to still refine probably some command line options and
configuration around this command so that people could have it fetch
prior to showing if they want. That way it is always up to date for
them. But they should also be able to override this and when this
command is paired with other commands like say as a prefix to the pull
command we should be able to override and control that configuration.

ps-id: 86f2b221-7805-458a-9d7c-81fba6d74c6d