Switch gps pull over to using gps fetch
Switch gps pull over to using gps fetch so that we could have it output
the upstream patches after doing the fetch. This is useful as it can
clue you into the changes that you just pulled down and rebased your
stack onto.

This is good to just be aware of changes as they are coming into the
codebase and how they might relate to what you are working on, or how
they might relate to mentoring you are doing, etc.

ps-id: 861e502c-4bc7-48ea-bfde-7f377bad0afd