Improve the output of the pull command
Improve the output of the pull command so that people can have a better
understanding of what operations it is performing for them and what
output belongs to which operation. I also added a couple blank lines of
output to create visual separation between the sections.

[changelog]
changed: output of pull command to clearly show commands and their output

ps-id: 879427bf-bfcf-4378-8e89-e02b80987085