Add isolate_post_cleanup hook execution to isolate
Add isolate_post_cleanup hook execution to isolate so that when it is
done with it's cleanup, checking out the branch you were on prior to
running isolate and deleting the ps/tmp/isolate branch, it will then run
this isolate_post_cleanup hook.

This will make it so that users can use this to support making their
package situation match the package systems associated lock file when it
switches back to their stack.

[changelog]
added: isolate_post_cleanup hook execution to the isolate command

ps-id: a75f2801-3e3f-4d06-86a8-f3a689abd2f9