Add util to trim new lines in string
Add a utility function to take a string and trim a new lines in it, if it has any. This will be used to format the output of the list_additional_info hook, since a new lines would break the layout of the whole list.

I created a function that borrows a string slice and returns a string without new line characters. I also import the module in mod.rs so that the test is picked up.

Relates to uptech/git-ps-rs#123

ps-id: d73136f2-080b-4fde-9d37-940dfbd153d5