Add execute with output function
Add a function to execute a script and return it's output, so we can use it with the list_additional_info hook.

I added a method to the execute module. The function accepts a file name to execute and returns the output on success.
I chose to add the function to the execute file instead of creating another file because it is very related to the execute util and it makes sense to have them together.

Relates to uptech/git-ps-rs#123

ps-id: d851dd86-a362-4d44-a81a-783f071b3f6a