Add integrate_post_push hook support
Add integrate_post_push hook support so that there is a lifecycle hook
for the integration process that gets fired after successfully
integrating the patch.

This will specifically be useful in a case where we are implementing a
code coverage storage mechanism tied to integrated commit shas so that
we can later grab the code coverage and compare patch specific code
coverage against it in the isolation verification step. Having this hook
will allow us to do this only in the scenarios where it successfully
integrated a patch.

[changelog]
added: integrate_post_push hook support

ps-id: 9edacf6c-1235-4d35-9a6a-c294d4b5d507