Add integrated commit's sha as arg to integrate_post_push hook
Add integrated commit's sha as arg to integrate_post_push hook so that
the hook can be aware of what commit was just successfully integrated.

To accomplish this I first modified the request_review_branch()
function to return the sha of the new commit created by cherry-picking.
This then enabled me to use this to pass it to the hook as an argument.

ps-id: 78fb1761-a97d-4646-b4aa-e60ace7b5c0c