Fix bug where int was referencing an old ps base
I in the integrate command it was grabbing the patch stack base right at
the beginning of the command. This is a problem because later on in the
command it does a fetch and then after that fetch it does a comparison
to determine of the patch is behind. It does this using the patch stack
base it obtained at the very beginning which is NOT updated by the
fetch. Therefore it is possible to be in a situation where even though
the patch is behind gps doesn't think it is. This would then allow it to
run the isolation verification process and then try to do the fast
forward push and fail.

The fix for this is to simply move the obtaining of the patch stack base
after we do the fetch that way it is as up to date as it possibly can
be so that the check for the patch being behind is actually valid.

This relates to issue #155

[changelog]
fixed: bug causing integrate patch behind check to not always be correct

ps-id: a2b0fe75-cf02-4b47-a0c9-8adb710c3aa6