Clean up duplication of merging optional mergables
We had a scenario were we were effectively duplicating the logic for
handling the merging of optional mergables. So I extracted that common
logic into a function called merge_option that I included in thhe
mergable.rs.

Then I used that function to replace all the instances where we were
duplicating the logic. I think this also helps to make the intention of
teh code clearer.

ps-id: 44f5fe1c-64d4-4d3d-8bb1-b782afeb34fb