Get additional information from hook output
Add a function that executes the list_additional_information hook and returns the formatted output of it.

I added a function in the private folder that calls `execute_with_output` on the hook and returns a string that is 10 characters long without a trailing new line, which would break the output.

Relates to uptech/git-ps-rs#123

ps-id: 66cbc199-1918-4733-83e2-dbd6291d7d7c