Refactor hook output function
Refactor the hook output function to return the output object instead of a string, so that the consumer of the function can decide what to do with it.

Also renamed the function to be a bit more descriptive of its function, and removed the part of finding the hook path since it was a bit redundant.

Relates to #123

ps-id: 85758df1-f9a1-4a32-b9c2-c0145c4e4f56