Add option to reverse order of stack in list cmd
Add option to reverse order of stack in list command so users that
prefer to see the stack in an order that matches the order of the
command line driven git rebase.

To enable this a user would simply need to add the following to their
config.

```
[list]
reverse_order = true
```

Signed-off-by: Ali Caglayan <alizter@gmail.com>

ps-id: 8f3dbfb9-46c6-43c5-be11-cbc36b05d25f

[changelog]
added: option to reverse patch stack order