Change add_ps_id() to use common anccestor
Change add_ps_id() to use common anccestor of the current branch and the
current branch's upstream branch.

I did this because add_ps_id() was rebasing from the base (origin/main)
up to head (main). This is fine in most cases. However, there is a
scenario where because of a git fetch that has occured origin/main has
been moved from where the local stack was actually based. When in this
case and the add_ps_id() process would happen it would effectively
rebase the stack on top of the newly updated origin/main and leave the
inverse of the patch(es) that were in between the orginal stack base
that origin/main used to point to and the newly updated origin/main as
local uncommitted changes.

To reproduce this scenario I did the following.

1. createed a bare repository locally
  - `mkdir gps_test_upstream_repo_bare`
  - `cd gps_test_upstream_repo_bare'
  - `git init --bare .`
  - `cd ..`
2. clone the bare repo to a repo to represent dev_a's repo
  - `git clone gps_test_upstream_repo_bare gps_test_dev_a_repo`
3. dev_a add initial commit to repo
  - `cd gps_test_dev_a_repo`
  - `echo "README" > README.md`
  - `git add README.md`
  - `git commit -m 'Initial commit'`
  - `git push`
4. dev_b clone the bare repo to a repo to represent dev_b's repo
  - `git clone gps_test_upstream_repo_bare gps_test_dev_b_repo`

At this point dev_a and dev_b's Git trees are the same and look as
follows.

```
* 781836e - (HEAD -> main, origin/main, origin/HEAD) Initial commit
```

5. dev_b locally creates a commit
  - `echo "1" > 1.txt`
  - `git add 1.txt`
  - `git commit -m 'add 1.txt'`

At this point dev_b's Git tree looks as follows.

```
* 098c04e - (HEAD -> main) add 1.txt
* 781836e - (origin/main, origin/HEAD) Initial commit
```

6. dev_a locally creates a commit
  - `echo "2" > 2.txt`
  - `git add 2.txt`
  - `git commit -m 'add 2.txt'`

At this point dev_a's Git tree looks as follows.

```
* e31d8d4 - (HEAD -> main) add 2.txt
* 781836e - (origin/main) Initial commit
```

And remember at this point the bare repo's Git tree looks as follows.

```
* 781836e - (origin/main) Initial commit
```

7. dev_a pushes the `add 2.txt` commit up
  - `git push`

Leaving both dev_a's repo and the bare repo with the follow Git tree.

```
* e31d8d4 - (HEAD -> main, origin/main) add 2.txt
* 781836e - Initial commit
```

So at this point dev_a's repo and the bare repo are both aware of the
`add 2.txt` commit. However, dev_b's repository is not aware of it.

8. dev_b integrates his `add 1.txt` patch
  - `gps int -f 0`

Prior to running this command dev_b`s repo tree was.

```
* 098c04e - (HEAD -> main) add 1.txt
* 781836e - (origin/main, origin/HEAD) Initial commit
```

After running this command dev_b's repo tree looks as follows and there
are local uncommitted changes that are the inverse of the `add 2.txt`
commit. So uncommitted changes in this case that delete the `2.txt`
file.

```
* d6b877c - (origin/main, origin/HEAD) add 1.txt
| * 05c56e6 - (HEAD -> main) add 1.txt
|/
* e31d8d4 - add 2.txt
* 781836e - Initial commit
```

The Git tree here is incorrect because add_ps_id() has taken the `add
1.txt` commit that `HEAD` points to and lifted it up and replayed it on
top of the `add 2.txt` commit. The tree instead should look something
like the following and there should be no uncommitted changes introduced.

```
* d6b877c - (origin/main, origin/HEAD) add 1.txt
* e31d8d4 - add 2.txt
| * 05c56e6 - (HEAD -> main) add 1.txt
|/
* 781836e - Initial commit
```

This behavior is unexpected and isnt' actually corrrect. It is happening
because add_ps_id() is doing the rebasing based on origin/main after the
integrate command does a fetch which then moves origin/main to point to
the `add 2.txt` commit.

The correct behavior for add_ps_id() is to actually add the patch stack
id but not based on origin/main which may have been recently updated but
instead to do it based on the common ancestor of main and origin/main.

This prevents this inadventent rebasing of the stack onto the latest
origin/main from happening and also prevents the incorrect side effect
of producing local changes that are the inversions of the skipped
commits.

The approach to correct this was simply to change the add_ps_id()
process to get the common anccestor and create it's temporary branch on
top of that common anccestor. This is correct in this scenario and I
believe it should fine in the other scenarios as well.

I manually tested the following scenarios.

- another dev has integrated 1 change into main between your
  current state and your integrate
- another dev(s) have integrated 2 or more changes into main between
  your current state and your integrate
- nobody has integrated any changes into main between your current state
  and your integrate

[changelog]
fixed: bug where add_ps_id() was incorrectly rebasing patches ontop of updated stack bases (e.g. origin/main)