Add create remote branch functionality to branch command
Add create remote branch functionality to branch command so that users
would have the ease of use of having a utility where they could create
and push a branch from either a patch or a series of patches. This is
extremely useful when you are stuck working with a team that doesn't want
to do Git Patch Stack, but you do want to do Git Patch Stack.

Note: The main work for this was done by Ali Caglayan (@Alizter) via the
https://github.com/uptech/git-ps-rs/pull/201 However, the diff was more
complicated than it should have been due to formatting differences.
Therefore, I manually cleaned up Ali's work and turned it into these
changes.

This resolves issue #196

[changelog]
added: push to remote feature to branch command

<!-- ps-id: 07c4a201-e6fc-4c00-8647-5cf2abe8111c -->