Add branch.push_to_remote config option
Add branch.push_to_remote config option so that users could configure
gps to push to the remote automatically any time they do a gps branch.
This is etremely useful when you are using gps locally but the team you
are working with doesn't want to use gps. I made this option default to
false.

In addition to adding this option I also renamed the command line option
to control this as well so that it matches in naming and concept.

[changelog]
added: branch.push_to_remote configuration option

<!-- ps-id: ce01645a-cecc-4103-bef2-b864e01f4730 -->