Add support for git submodules
Add support for git submodules and simply make gps more correct in terms
of its pathing for future growth with git.

Prior to this patch there existed an issue where there was an assumption
implementing in some of the pathing within gps that was causing
problems. That assumption was that the .git directory would always be
inside of the working directory.

This assumption was impacting pathing for accessing the following

- gps config
- gps hooks
- state management

This assumption was an incorrect assumption. For instance if the user is
using git submodules the .git isn't even a directory. It is instead a
file that then has a relative path stored inside it to access to actual
.git directory.

This patch removes this assumption by explicitly depending on the .git
directory path from git2 and libgit2 in turn. That way it can be managed
in one place and if git decides to introduce another scenario where it
is different, gps should continue to work.

My approach to resolving this was to formally pass in the .git dir where
needed and then to update all the commands to appropriately get the .git
dir and pass it into the respective functions.

[changelog]
added: git submodule support

<!-- ps-id: 7abd7892-85c5-4cc3-b4e5-d7ec0c8c64cc -->