Add prompt for reassurance to int -f
Add prompt for reassurance to int -f so that the behavior is as expected
and not confusing when you use the -f/--force switch with gps int.

The approach was simply to add the prompt reassurance check and
execution to the force code path.

This addresses issue #217

[changelog]
added: prompt for reassurance to gps int -f

<!-- ps-id: 3faa5bf9-5bb7-4091-9a8c-d97f6b2e6ded -->