Use homdir to allow cross-platform support
Switch from the home-dir dependency for figuring out the users home
directory to the homedir dependency as it is a cross-platform library.
This should enable us to get the users home directory on unix based
platforms as well as windows.

This relates to issue #218.

[changelog]
changed: home-dir dep to homedir for cross-platform

Signed-off-by: Ali Caglayan <alizter@gmail.com>

<!-- ps-id: 4e02c031-268a-454a-b18f-81aac1cf199d -->