Switch from stored cached state to computed state
The intention of this change is to fundamentally change gps so that it
no longer stores and manages, local cache about things. Instead it
should now simply read the Git tree and compute state from that.

This should give us a lot more stability, and remove any brittleness
that existed between users switching back and forth between Git Patch
Stack commands and classic Git commands.

It should also have the benefit of allowing us to expose branches
associated with patches in a way to users that can help formalize the
concept of a series of patches and bring it forth as a first class
citizen.

It will have some impacts. For example we will no longer be able to
internally have and represent the state of if a patch has had a review
requested. This is actually good in my mind as it was state that is
actually owned by GitHub/Bitbucket/GitLab and we were simply cacheing it
before. To address this in the new world order we can use the
list_additional_info hook as part of the list command to fetch that
state from its natural source and include it in the gps ls output. We
also used that state with the request_review_post_sync hook before. But
now we can fetch that state from its natural source within the hook,
rather than getting handed the state from the cache.

We will also lose the ability know how patches changed since an
operation. For example previously we tracked the hash of a patches diff
when we performed an operation like syncing it to a remote. This enabled
us to determine if the patch was locally changed since the last sync,
remotely changed since the last sync, or both. We will no longer have
this state. Therefore all we will be able to determine is if the remote
patch and the local patch match or don't match. This is something that
we will just have to do without. But, I think it is worth the loss to
gain the stability and the formal patch series concept.

<!-- ps-id: be9e5f1f-75a7-4fce-8f8c-60c2a5d2254e -->

[changelog]
added: stack and following context to the ls cmd
added: showing associated branches to the ls cmd
changed: ls cmd to show branch associated patch states
changed: ls cmd to use computed state from git tree rather than tracked state
changed: request_review_branch cmd to use computed state rather than tracked state
changed: request_review cmd to use computed state rather than tracked state
changed: request_review_post_sync to no longer receive re-requesting review state
changed: request_review_post_sync examples to deal without re-requesting review
changed: sync cmd to use computed state rather than tracked state
changed: integrate cmd to use computed state rather than tracked state
changed: branch cmd to set upstream tracking branch when pushing